pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "gsheet-pandas"
version = "0.2.7"
description = "Download and upload pandas dataframes to the Google sheets"
authors = [{name = "Iakov Kaiumov", email = "yashahelfi2012@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.101.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
pandas = "^2.1.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
